Move IoTConsensus multi-data-dir region migration ITs to commit suite
