+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/regionmigration/pass/commit/IoTDBRegionMigrateWithDeletionMultiDataDirIT.java
@@ -17,15 +17,13 @@
  * under the License.
  */
 
-package org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1;
+package org.apache.iotdb.confignode.it.regionmigration.pass.commit;
 
 import org.apache.iotdb.consensus.ConsensusFactory;
-import org.apache.iotdb.isession.SessionConfig;
 import org.apache.iotdb.it.env.EnvFactory;
 import org.apache.iotdb.it.env.cluster.node.DataNodeWrapper;
 import org.apache.iotdb.it.framework.IoTDBTestRunner;
-import org.apache.iotdb.itbase.category.TableClusterIT;
-import org.apache.iotdb.itbase.env.BaseEnv;
+import org.apache.iotdb.itbase.category.ClusterIT;
 
 import org.apache.tsfile.utils.Pair;
 import org.awaitility.Awaitility;
@@ -45,21 +43,23 @@ import java.util.concurrent.TimeUnit;
 
 import static org.apache.iotdb.confignode.it.regionmigration.IoTDBRegionOperationReliabilityITFramework.getAllDataNodes;
 import static org.apache.iotdb.confignode.it.regionmigration.IoTDBRegionOperationReliabilityITFramework.getDataRegionMapWithLeader;
-import static org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1.RegionMigrateFileAssertions.MULTI_DATA_DIRS;
-import static org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1.RegionMigrateFileAssertions.awaitModsVisibleOnReplicas;
-import static org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1.RegionMigrateFileAssertions.awaitRegionReplicas;
-import static org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1.RegionMigrateFileAssertions.awaitTsFileResourceVisibleOnReplicas;
-import static org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1.RegionMigrateFileAssertions.awaitTsFileVisibleOnReplicas;
-import static org.apache.iotdb.confignode.it.regionmigration.pass.daily.iotv1.RegionMigrateFileAssertions.getReplicaDataNodeIds;
+import static org.apache.iotdb.confignode.it.regionmigration.pass.commit.RegionMigrateFileAssertions.MULTI_DATA_DIRS;
+import static org.apache.iotdb.confignode.it.regionmigration.pass.commit.RegionMigrateFileAssertions.awaitModsVisibleOnReplicas;
+import static org.apache.iotdb.confignode.it.regionmigration.pass.commit.RegionMigrateFileAssertions.awaitRegionReplicas;
+import static org.apache.iotdb.confignode.it.regionmigration.pass.commit.RegionMigrateFileAssertions.awaitTsFileResourceVisibleOnReplicas;
+import static org.apache.iotdb.confignode.it.regionmigration.pass.commit.RegionMigrateFileAssertions.awaitTsFileVisibleOnReplicas;
+import static org.apache.iotdb.confignode.it.regionmigration.pass.commit.RegionMigrateFileAssertions.getReplicaDataNodeIds;
 
 /**
- * Table-model twin of {@link IoTDBRegionMigrateWithDeletionMultiDataDirIT}: a deletion (mods) must
- * survive IoTConsensus region migration across multiple data dirs, asserted through the relational
- * (table) SQL dialect so the table-model cluster CI covers the same snapshot mods-transfer path.
+ * Tree-model coverage for IoTConsensus region migration over multiple data dirs: a deletion (mods)
+ * must survive the snapshot transfer to the migrated peer. With several data dirs the snapshot
+ * fragments of one TsFile can be received into different folders, so the receiver groups companion
+ * files and the loader relinks them into one data dir; if that breaks, the migrated replica loses
+ * the deletion. See the table-model twin {@link IoTDBRegionMigrateWithDeletionMultiDataDirTableIT}.
  */
 @RunWith(IoTDBTestRunner.class)
-@Category({TableClusterIT.class})
-public class IoTDBRegionMigrateWithDeletionMultiDataDirTableIT {
+@Category({ClusterIT.class})
+public class IoTDBRegionMigrateWithDeletionMultiDataDirIT {
 
   @Before
   public void setUp() throws Exception {
@@ -79,12 +79,11 @@ public class IoTDBRegionMigrateWithDeletionMultiDataDirTableIT {
 
   @Test
   public void testRegionMigratePreservesDeletionWithMultiDataDirs() throws Exception {
-    try (Connection connection = EnvFactory.getEnv().getTableConnection();
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      statement.execute("CREATE DATABASE test");
-      statement.execute("USE test");
-      statement.execute("CREATE TABLE t1 (s1 INT64 FIELD)");
-      statement.execute("INSERT INTO t1 (time, s1) VALUES (100, 100), (200, 200), (300, 300)");
+      statement.execute("CREATE DATABASE root.db");
+      statement.execute(
+          "INSERT INTO root.db.d1(timestamp, s1) VALUES (100, 100), (200, 200), (300, 300)");
       statement.execute("FLUSH");
 
       Map<Integer, Pair<Integer, Set<Integer>>> dataRegionMapWithLeader =
@@ -94,14 +93,14 @@ public class IoTDBRegionMigrateWithDeletionMultiDataDirTableIT {
       Set<Integer> initialReplicaDataNodeIds =
           getReplicaDataNodeIds(statement, dataRegionIdForTest);
 
-      awaitTsFileVisibleOnReplicas("test", dataRegionIdForTest, initialReplicaDataNodeIds);
+      awaitTsFileVisibleOnReplicas("root.db", dataRegionIdForTest, initialReplicaDataNodeIds);
       awaitTsFileResourceVisibleOnReplicas(
-          statement, "test", dataRegionIdForTest, initialReplicaDataNodeIds);
+          statement, "root.db", dataRegionIdForTest, initialReplicaDataNodeIds);
 
-      statement.execute("DELETE FROM t1 WHERE time <= 200");
+      statement.execute("DELETE FROM root.db.d1.s1 WHERE time <= 200");
       statement.execute("FLUSH");
-      awaitModsVisibleOnReplicas("test", dataRegionIdForTest, initialReplicaDataNodeIds);
-      assertDeletionVisibleOnAllReplicas(statement, dataRegionIdForTest, 1);
+      awaitModsVisibleOnReplicas("root.db", dataRegionIdForTest, initialReplicaDataNodeIds);
+      assertDeletionVisibleOnAllReplicas(dataRegionIdForTest, 1);
 
       Pair<Integer, Set<Integer>> leaderAndNodes = dataRegionMapWithLeader.get(dataRegionIdForTest);
       Set<Integer> allDataNodes = getAllDataNodes(statement);
@@ -119,15 +118,19 @@ public class IoTDBRegionMigrateWithDeletionMultiDataDirTableIT {
               "migrate region %d from %d to %d", dataRegionIdForTest, leaderId, destDataNodeId));
 
       awaitRegionReplicas(statement, dataRegionIdForTest, Set.of(followerId, destDataNodeId));
-      awaitModsVisibleOnReplicas("test", dataRegionIdForTest, Set.of(destDataNodeId));
+      awaitModsVisibleOnReplicas("root.db", dataRegionIdForTest, Set.of(destDataNodeId));
 
-      assertDeletionVisibleOnAllReplicas(statement, dataRegionIdForTest, 1);
+      assertDeletionVisibleOnAllReplicas(dataRegionIdForTest, 1);
     }
   }
 
-  private void assertDeletionVisibleOnAllReplicas(
-      Statement statement, int dataRegionId, int expectedCount) throws Exception {
-    Set<Integer> replicaDataNodeIds = getReplicaDataNodeIds(statement, dataRegionId);
+  private void assertDeletionVisibleOnAllReplicas(int dataRegionId, int expectedCount)
+      throws Exception {
+    Set<Integer> replicaDataNodeIds;
+    try (Connection connection = EnvFactory.getEnv().getConnection();
+        Statement statement = connection.createStatement()) {
+      replicaDataNodeIds = getReplicaDataNodeIds(statement, dataRegionId);
+    }
     for (int dataNodeId : replicaDataNodeIds) {
       DataNodeWrapper dataNodeWrapper =
           EnvFactory.getEnv().dataNodeIdToWrapper(dataNodeId).orElseThrow();
@@ -141,21 +144,15 @@ public class IoTDBRegionMigrateWithDeletionMultiDataDirTableIT {
 
   private void assertDeletionVisibleOnReplica(DataNodeWrapper dataNodeWrapper, int expectedCount)
       throws Exception {
-    try (Connection connection =
-            EnvFactory.getEnv()
-                .getConnection(
-                    dataNodeWrapper,
-                    SessionConfig.DEFAULT_USER,
-                    SessionConfig.DEFAULT_PASSWORD,
-                    BaseEnv.TABLE_SQL_DIALECT);
+    try (Connection connection = EnvFactory.getEnv().getConnection(dataNodeWrapper);
         Statement dataNodeStatement = connection.createStatement()) {
-      dataNodeStatement.execute("USE test");
-      try (ResultSet countResultSet = dataNodeStatement.executeQuery("SELECT COUNT(s1) FROM t1")) {
+      try (ResultSet countResultSet =
+          dataNodeStatement.executeQuery("SELECT COUNT(s1) FROM root.db.d1")) {
         Assert.assertTrue(countResultSet.next());
         Assert.assertEquals(expectedCount, countResultSet.getLong(1));
       }
       try (ResultSet deletedRangeResultSet =
-          dataNodeStatement.executeQuery("SELECT s1 FROM t1 WHERE time <= 200")) {
+          dataNodeStatement.executeQuery("SELECT s1 FROM root.db.d1 WHERE time <= 200")) {
         Assert.assertFalse(deletedRangeResultSet.next());
       }
     }
